refactor(session-store): extract session-cold + reference index helpers (file #4)
Lift the session-list cold-data array/index helpers + the session reference
projections (live-sync + palette) and their types into session-cold-index.ts;
session-store re-exports the reference types. Pure list transforms. check 0,
store vitest parity (503 pass).
