+++ packages/desktop/src/lib/acp/store/session-store.svelte.ts
@@ -64,6 +64,24 @@ import {
 	seedTranscriptEntryIndex,
 } from "./transcript-entry-index.js";
 export { transcriptEntryIndexes } from "./transcript-entry-index.js";
+import {
+	createPatchedReferenceArray,
+	createPatchedSessionColdArray,
+	createPrependedReferenceArray,
+	createPrependedSessionColdArray,
+	findSessionColdIndexById,
+	patchSessionIdsByProjectIndex,
+	patchSessionsByProjectIndex,
+	rebuildLiveSessionSyncReferences,
+	rebuildSessionByIdIndex,
+	rebuildSessionIdsByProjectIndex,
+	rebuildSessionPaletteReferences,
+	rebuildSessionsByProjectIndex,
+	sessionLiveSyncReferenceFromSession,
+	sessionPaletteReferenceFromSession,
+} from "./session-cold-index.js";
+import type { SessionLiveSyncReference, SessionPaletteReference } from "./session-cold-index.js";
+export type { SessionLiveSyncReference, SessionPaletteReference } from "./session-cold-index.js";
 export {
 	mergeInteractionSnapshots,
 	mergeOperationSnapshots,
@@ -226,17 +244,6 @@ function isSessionQuestionInteraction(
 
 type InflightSessionStateRefresh = ResultAsync<void, AppError>;
 
-export type SessionLiveSyncReference = {
-	readonly id: string;
-	readonly updatedAtMs: number;
-};
-
-export type SessionPaletteReference = {
-	readonly id: string;
-	readonly projectPath: string;
-	readonly agentId: string;
-	readonly title: string | null;
-};
 
 
 
@@ -504,426 +511,6 @@ function graphWithPatches(input: {
 		activity: input.activity ?? input.graph.activity,
 		capabilities: input.graph.capabilities,
 	};
-}
-
-function createPrependedSessionColdArray(
-	session: SessionCold,
-	base: readonly SessionCold[]
-): SessionCold[] {
-	const target = new Array<SessionCold>(base.length + 1);
-	return new Proxy(target, {
-		get(targetArray, property, receiver) {
-			if (property === Symbol.iterator) {
-				return function* () {
-					for (let index = 0; index < targetArray.length; index += 1) {
-						yield selectPrependedSessionCold(session, base, index);
-					}
-				};
-			}
-			if (typeof property === "string") {
-				const index = toArrayIndex(property);
-				if (index !== null) {
-					return selectPrependedSessionCold(session, base, index);
-				}
-				if (property === "slice") {
-					return (start?: number, end?: number) =>
-						Array.prototype.slice.call(receiver, start, end);
-				}
-			}
-			const value = Reflect.get(targetArray, property, receiver);
-			return typeof value === "function" ? value.bind(receiver) : value;
-		},
-		has(targetArray, property) {
-			const index = typeof property === "string" ? toArrayIndex(property) : null;
-			if (index !== null) {
-				return index >= 0 && index < targetArray.length;
-			}
-			return property in targetArray;
-		},
-		getOwnPropertyDescriptor(targetArray, property) {
-			const index = typeof property === "string" ? toArrayIndex(property) : null;
-			if (index !== null && index >= 0 && index < targetArray.length) {
-				return {
-					configurable: true,
-					enumerable: true,
-					value: selectPrependedSessionCold(session, base, index),
-					writable: false,
-				};
-			}
-			return Reflect.getOwnPropertyDescriptor(targetArray, property);
-		},
-	});
-}
-
-function selectPrependedSessionCold(
-	session: SessionCold,
-	base: readonly SessionCold[],
-	index: number
-): SessionCold | undefined {
-	return index === 0 ? session : base[index - 1];
-}
-
-function createPatchedSessionColdArray(
-	base: readonly SessionCold[],
-	patchedIndex: number,
-	session: SessionCold
-): SessionCold[] {
-	const target = new Array<SessionCold>(base.length);
-	return new Proxy(target, {
-		get(targetArray, property, receiver) {
-			if (property === Symbol.iterator) {
-				return function* () {
-					for (let index = 0; index < targetArray.length; index += 1) {
-						yield index === patchedIndex ? session : base[index];
-					}
-				};
-			}
-			if (typeof property === "string") {
-				const index = toArrayIndex(property);
-				if (index !== null) {
-					return index === patchedIndex ? session : base[index];
-				}
-				if (property === "slice") {
-					return (start?: number, end?: number) =>
-						Array.prototype.slice.call(receiver, start, end);
-				}
-			}
-			const value = Reflect.get(targetArray, property, receiver);
-			return typeof value === "function" ? value.bind(receiver) : value;
-		},
-		has(targetArray, property) {
-			const index = typeof property === "string" ? toArrayIndex(property) : null;
-			if (index !== null) {
-				return index >= 0 && index < targetArray.length;
-			}
-			return property in targetArray;
-		},
-		getOwnPropertyDescriptor(targetArray, property) {
-			const index = typeof property === "string" ? toArrayIndex(property) : null;
-			if (index !== null && index >= 0 && index < targetArray.length) {
-				return {
-					configurable: true,
-					enumerable: true,
-					value: index === patchedIndex ? session : base[index],
-					writable: false,
-				};
-			}
-			return Reflect.getOwnPropertyDescriptor(targetArray, property);
-		},
-	});
-}
-
-function findSessionColdIndexById(sessions: readonly SessionCold[], sessionId: string): number {
-	for (let index = 0; index < sessions.length; index += 1) {
-		if (sessions[index]?.id === sessionId) {
-			return index;
-		}
-	}
-	return -1;
-}
-
-function rebuildSessionByIdIndex(
-	index: SvelteMap<string, SessionCold>,
-	sessions: readonly SessionCold[]
-): void {
-	index.clear();
-	for (const session of sessions) {
-		index.set(session.id, session);
-	}
-}
-
-function rebuildSessionsByProjectIndex(
-	index: SvelteMap<string, SessionCold[]>,
-	sessions: readonly SessionCold[]
-): void {
-	index.clear();
-	for (const session of sessions) {
-		const projectSessions = index.get(session.projectPath);
-		if (projectSessions === undefined) {
-			index.set(session.projectPath, [session]);
-			continue;
-		}
-		projectSessions.push(session);
-	}
-}
-
-function rebuildSessionIdsByProjectIndex(
-	index: SvelteMap<string, string[]>,
-	sessions: readonly SessionCold[]
-): void {
-	index.clear();
-	for (const session of sessions) {
-		const projectSessionIds = index.get(session.projectPath);
-		if (projectSessionIds === undefined) {
-			index.set(session.projectPath, [session.id]);
-			continue;
-		}
-		projectSessionIds.push(session.id);
-	}
-}
-
-function patchSessionsByProjectIndex(
-	index: SvelteMap<string, SessionCold[]>,
-	previousSession: SessionCold | undefined,
-	nextSession: SessionCold
-): void {
-	if (previousSession !== undefined && previousSession.projectPath !== nextSession.projectPath) {
-		const previousProjectSessions = index.get(previousSession.projectPath);
-		if (previousProjectSessions !== undefined) {
-			const nextPreviousProjectSessions = removeSessionColdFromArray(
-				previousProjectSessions,
-				previousSession.id
-			);
-			if (nextPreviousProjectSessions.length === 0) {
-				index.delete(previousSession.projectPath);
-			} else {
-				index.set(previousSession.projectPath, nextPreviousProjectSessions);
-			}
-		}
-	}
-
-	const currentProjectSessions = index.get(nextSession.projectPath);
-	if (currentProjectSessions === undefined) {
-		index.set(nextSession.projectPath, [nextSession]);
-		return;
-	}
-
-	const projectIndex = findSessionColdIndexById(currentProjectSessions, nextSession.id);
-	if (projectIndex === -1) {
-		index.set(
-			nextSession.projectPath,
-			createPrependedSessionColdArray(nextSession, currentProjectSessions)
-		);
-		return;
-	}
-
-	index.set(
-		nextSession.projectPath,
-		createPatchedSessionColdArray(currentProjectSessions, projectIndex, nextSession)
-	);
-}
-
-function patchSessionIdsByProjectIndex(
-	index: SvelteMap<string, string[]>,
-	previousSession: SessionCold | undefined,
-	nextSession: SessionCold
-): void {
-	if (previousSession !== undefined && previousSession.projectPath !== nextSession.projectPath) {
-		const previousProjectSessionIds = index.get(previousSession.projectPath);
-		if (previousProjectSessionIds !== undefined) {
-			const nextPreviousProjectSessionIds = removeSessionIdFromArray(
-				previousProjectSessionIds,
-				previousSession.id
-			);
-			if (nextPreviousProjectSessionIds.length === 0) {
-				index.delete(previousSession.projectPath);
-			} else {
-				index.set(previousSession.projectPath, nextPreviousProjectSessionIds);
-			}
-		}
-	}
-
-	const currentProjectSessionIds = index.get(nextSession.projectPath);
-	if (currentProjectSessionIds === undefined) {
-		index.set(nextSession.projectPath, [nextSession.id]);
-		return;
-	}
-
-	if (currentProjectSessionIds.includes(nextSession.id)) {
-		return;
-	}
-
-	index.set(
-		nextSession.projectPath,
-		createPrependedReferenceArray(nextSession.id, currentProjectSessionIds)
-	);
-}
-
-function removeSessionColdFromArray(
-	sessions: readonly SessionCold[],
-	sessionId: string
-): SessionCold[] {
-	const removedIndex = findSessionColdIndexById(sessions, sessionId);
-	if (removedIndex === -1) {
-		return sessions as SessionCold[];
-	}
-	const nextSessions: SessionCold[] = [];
-	for (let index = 0; index < sessions.length; index += 1) {
-		if (index !== removedIndex) {
-			const session = sessions[index];
-			if (session !== undefined) {
-				nextSessions.push(session);
-			}
-		}
-	}
-	return nextSessions;
-}
-
-function removeSessionIdFromArray(sessionIds: readonly string[], sessionId: string): string[] {
-	const removedIndex = sessionIds.indexOf(sessionId);
-	if (removedIndex === -1) {
-		return sessionIds as string[];
-	}
-	const nextSessionIds: string[] = [];
-	for (let index = 0; index < sessionIds.length; index += 1) {
-		if (index !== removedIndex) {
-			const existingSessionId = sessionIds[index];
-			if (existingSessionId !== undefined) {
-				nextSessionIds.push(existingSessionId);
-			}
-		}
-	}
-	return nextSessionIds;
-}
-
-function sessionLiveSyncReferenceFromSession(session: SessionCold): SessionLiveSyncReference {
-	return {
-		id: session.id,
-		updatedAtMs: session.updatedAt.getTime(),
-	};
-}
-
-function sessionPaletteReferenceFromSession(session: SessionCold): SessionPaletteReference {
-	return {
-		id: session.id,
-		projectPath: session.projectPath,
-		agentId: session.agentId,
-		title: session.title,
-	};
-}
-
-function createPrependedReferenceArray<TReference>(
-	reference: TReference,
-	base: readonly TReference[]
-): TReference[] {
-	const target = new Array<TReference>(base.length + 1);
-	return new Proxy(target, {
-		get(targetArray, property, receiver) {
-			if (property === Symbol.iterator) {
-				return function* () {
-					for (let index = 0; index < targetArray.length; index += 1) {
-						yield index === 0 ? reference : base[index - 1];
-					}
-				};
-			}
-			if (typeof property === "string") {
-				const index = toArrayIndex(property);
-				if (index !== null) {
-					return index === 0 ? reference : base[index - 1];
-				}
-				if (property === "slice") {
-					return (start?: number, end?: number) =>
-						Array.prototype.slice.call(receiver, start, end);
-				}
-			}
-			const value = Reflect.get(targetArray, property, receiver);
-			return typeof value === "function" ? value.bind(receiver) : value;
-		},
-		has(targetArray, property) {
-			const index = typeof property === "string" ? toArrayIndex(property) : null;
-			if (index !== null) {
-				return index >= 0 && index < targetArray.length;
-			}
-			return property in targetArray;
-		},
-		getOwnPropertyDescriptor(targetArray, property) {
-			const index = typeof property === "string" ? toArrayIndex(property) : null;
-			if (index !== null && index >= 0 && index < targetArray.length) {
-				return {
-					configurable: true,
-					enumerable: true,
-					value: index === 0 ? reference : base[index - 1],
-					writable: false,
-				};
-			}
-			return Reflect.getOwnPropertyDescriptor(targetArray, property);
-		},
-	});
-}
-
-function createPatchedReferenceArray<TReference extends { readonly id: string }>(
-	base: readonly TReference[],
-	reference: TReference
-): TReference[] {
-	const patchedIndex = findReferenceIndexById(base, reference.id);
-	if (patchedIndex === -1) {
-		return createPrependedReferenceArray(reference, base);
-	}
-	const target = new Array<TReference>(base.length);
-	return new Proxy(target, {
-		get(targetArray, property, receiver) {
-			if (property === Symbol.iterator) {
-				return function* () {
-					for (let index = 0; index < targetArray.length; index += 1) {
-						yield index === patchedIndex ? reference : base[index];
-					}
-				};
-			}
-			if (typeof property === "string") {
-				const index = toArrayIndex(property);
-				if (index !== null) {
-					return index === patchedIndex ? reference : base[index];
-				}
-				if (property === "slice") {
-					return (start?: number, end?: number) =>
-						Array.prototype.slice.call(receiver, start, end);
-				}
-			}
-			const value = Reflect.get(targetArray, property, receiver);
-			return typeof value === "function" ? value.bind(receiver) : value;
-		},
-		has(targetArray, property) {
-			const index = typeof property === "string" ? toArrayIndex(property) : null;
-			if (index !== null) {
-				return index >= 0 && index < targetArray.length;
-			}
-			return property in targetArray;
-		},
-		getOwnPropertyDescriptor(targetArray, property) {
-			const index = typeof property === "string" ? toArrayIndex(property) : null;
-			if (index !== null && index >= 0 && index < targetArray.length) {
-				return {
-					configurable: true,
-					enumerable: true,
-					value: index === patchedIndex ? reference : base[index],
-					writable: false,
-				};
-			}
-			return Reflect.getOwnPropertyDescriptor(targetArray, property);
-		},
-	});
-}
-
-function findReferenceIndexById<TReference extends { readonly id: string }>(
-	references: readonly TReference[],
-	id: string
-): number {
-	for (let index = 0; index < references.length; index += 1) {
-		if (references[index]?.id === id) {
-			return index;
-		}
-	}
-	return -1;
-}
-
-function rebuildLiveSessionSyncReferences(
-	sessions: readonly SessionCold[]
-): SessionLiveSyncReference[] {
-	const references: SessionLiveSyncReference[] = [];
-	for (const session of sessions) {
-		references.push(sessionLiveSyncReferenceFromSession(session));
-	}
-	return references;
-}
-
-function rebuildSessionPaletteReferences(
-	sessions: readonly SessionCold[]
-): SessionPaletteReference[] {
-	const references: SessionPaletteReference[] = [];
-	for (const session of sessions) {
-		references.push(sessionPaletteReferenceFromSession(session));
-	}
-	return references;
 }
 
 export function applyTranscriptDeltaToSnapshot(
